+++ tests/files/src/0716_false_positive_in_elseif.php
@@ -1,0 +1,8 @@
+<?php
+
+function test($x) {
+    if ($x > 0) {
+    } elseif ($x = (int)$x) {  // should not emit PhanRedundantCondition for cast to int
+        echo $x;
+    }
+}