+++ tests/files/src/0717_variable_assign_in_complex_condition.php
@@ -1,0 +1,13 @@
+<?php
+class Example {
+    /** @return array */
+    public function read($token)
+    {
+        if (!$token || !file_exists($file = 'somefile.txt')) {
+            return [];
+        }
+
+        // Should infer the type of $file
+        return $file;
+    }
+}