+++ NEWS.md
@@ -4,6 +4,9 @@ Jun 17 2019, Phan 2.2.4 (dev)
 -----------------------
 
 + Reduce false positives inferring the resulting type of `$x++`, `--$x`, etc. (#2877)
++ Fix false positives analyzing variable modification in `elseif` conditions (#2878, #2860)
+  (e.g. no longer emit `PhanRedundantCondition` analyzing `elseif ($offset = (int)$offset)`)
+  (e.g. do a better job inferring variables set in complex `if` condition expressions)
 
 Jun 17 2019, Phan 2.2.3
 -----------------------