+++ NEWS.md
@@ -3,6 +3,10 @@ Phan NEWS
 Jun 17 2019, Phan 2.2.4 (dev)
 -----------------------
 
+New features(CLI, Configs):
++ Warn if any of the files passed in `--include-analysis-file-list` don't exist.
+
+New features(Analysis):
 + Reduce false positives inferring the resulting type of `$x++`, `--$x`, etc. (#2877)
 + Fix false positives analyzing variable modification in `elseif` conditions (#2878, #2860)
   (e.g. no longer emit `PhanRedundantCondition` analyzing `elseif ($offset = (int)$offset)`)